fix: exec commands are given directory cwd
BREAKING CHANGES:

If you had have used a `before` or `after` step - that was a string command,
then this used to be called with the `cwd` set to the filename of that file,
for example provisioning `package.json` would provide the `cwd` of
`/foo/bar/package.json`. This change means that the directory is passed as
the `cwd` - so the `cwd` would be set to `/foo/bar` in the above example.